perceptron/pyproject.toml
[project]
name = "perceptron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.4",
    "pillow>=11.2.1",
]
